feat: refactor firmware update components for improved structure and functionality
